requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Database
asyncpg==0.29.0
redis==5.0.1
sqlalchemy==2.0.23

# HTTP Client
aiohttp==3.9.1
httpx==0.25.2

# RSS/Feed Processing
feedparser==6.0.10

# Payment Processing
stripe==7.0.0

# Email
sendgrid==6.11.0

# Monitoring & Observability
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-exporter-otlp-proto-grpc==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-instrumentation-asyncpg==0.42b0

# Utilities
python-dateutil==2.8.2
pytz==2023.3
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-playwright==0.4.3

# Code Quality
ruff==0.1.9  # Replaces flake8, black, and isort
mypy==1.7.0

# Type Stubs
types-redis==4.6.0.11
types-python-dateutil==2.8.19.14

# Aurora Prime Dependencies
httpx==0.25.2
pyyaml==6.0.1
